move ChannelMessageRegistrator to Channel/ChannelMessageRegistrator
