+++ worker/include/Shared.hpp
@@ -1,8 +1,8 @@
 #ifndef MS_SHARED_HPP
 #define MS_SHARED_HPP
 
-#include "ChannelMessageRegistrator.hpp"
 #include "SharedInterface.hpp"
+#include "Channel/ChannelMessageRegistrator.hpp"
 #include "Channel/ChannelNotifier.hpp"
 #include "handles/BackoffTimerHandleInterface.hpp"
 #include "handles/TimerHandleInterface.hpp"
@@ -11,12 +11,13 @@ class Shared : public SharedInterface
 {
 public:
 	explicit Shared(
-	  ChannelMessageRegistrator* channelMessageRegistrator, Channel::ChannelNotifier* channelNotifier);
+	  Channel::ChannelMessageRegistrator* channelMessageRegistrator,
+	  Channel::ChannelNotifier* channelNotifier);
 
 	~Shared() override;
 
 public:
-	ChannelMessageRegistrator* GetChannelMessageRegistrator() const override
+	Channel::ChannelMessageRegistrator* GetChannelMessageRegistrator() const override
 	{
 		return this->channelMessageRegistrator.get();
 	}
@@ -32,7 +33,7 @@ public:
 	  const BackoffTimerHandleInterface::BackoffTimerHandleOptions& options) const override;
 
 private:
-	std::unique_ptr<ChannelMessageRegistrator> channelMessageRegistrator;
+	std::unique_ptr<Channel::ChannelMessageRegistrator> channelMessageRegistrator;
 	std::unique_ptr<Channel::ChannelNotifier> channelNotifier;
 };
