+++ worker/meson.build
@@ -95,7 +95,6 @@ common_sources = [
   'src/Settings.cpp',
   'src/Shared.cpp',
   'src/Worker.cpp',
-  'src/ChannelMessageRegistrator.cpp',
   'src/Utils/BitStream.cpp',
   'src/Utils/Crypto.cpp',
   'src/Utils/File.cpp',
@@ -110,6 +109,7 @@ common_sources = [
   'src/handles/UnixStreamSocketHandle.cpp',
   'src/Channel/ChannelNotifier.cpp',
   'src/Channel/ChannelRequest.cpp',
+  'src/Channel/ChannelMessageRegistrator.cpp',
   'src/Channel/ChannelNotification.cpp',
   'src/Channel/ChannelSocket.cpp',
   'src/RTC/ActiveSpeakerObserver.cpp',
